refactor(web): card-free stat row + extract PrivateProfileCard
