+++ apps/web/src/components/members/logged-out-teaser.tsx
@@ -6,12 +6,11 @@
 // through Clerk back to this page (redirect_url).
 
 import { MemberIdentityHeader } from "@/components/members/member-identity-header";
+import { PrivateProfileCard } from "@/components/members/private-profile-card";
 import { Button } from "@/components/ui/button";
-import { Card } from "@/components/ui/card";
 import { gateway } from "@/gateway";
 import { Link } from "@tanstack/react-router";
 import type { FunctionReturnType } from "convex/server";
-import { Lock } from "lucide-react";
 import { useTranslations } from "use-intl";
 
 // The web tier derives Convex view types from the gateway (not @jigswap/contracts directly).
@@ -68,21 +67,5 @@ export function LoggedOutTeaser({
         </Button>
       </div>
     </div>
-  );
-}
-
-// The quiet private-profile card (shared shape with the logged-in interstitial):
-// full identity above, one sentence of mutuality framing, no blurred-content
-// silhouettes — respectful, never a paywall tease.
-export function PrivateProfileCard({ displayName }: { displayName: string }) {
-  const t = useTranslations("members");
-  return (
-    <Card className="flex flex-col items-center gap-2 border-dashed p-8 text-center">
-      <Lock className="text-muted-foreground h-5 w-5" />
-      <p className="font-semibold">
-        {t("privateTitle", { name: displayName })}
-      </p>
-      <p className="text-muted-foreground text-sm">{t("privateSub")}</p>
-    </Card>
   );
 }
